0.2 - kinder surprise enemy, powerups, safer enemy spawns

+++ main.js
@@ -1,17 +1,17 @@
-window.onerror = function(e, url, line) {
+/*window.onerror = function(e, url, line) {
     alert(e+'\n'+line+'\n'+url);
-}
-
+}*/
+var VERSION = '0.2.1';
 var KEYCODE_A = 65;
 var KEYCODE_S = 83;
 var KEYCODE_D = 68;
 var KEYCODE_W = 87;
+var GAME_WIDTH = 800;
+var GAME_HEIGHT = 450;
 
 var posneg = function() {
     return game.rnd.integerInRange(0,10) > 5 ? -1 : 1;
 };
-
-
 
 /*** Enemy Definitions! ***/
 
@@ -56,32 +56,112 @@ enemy1_controller.spritelist = ['enemy1', 'enemy2'];
 enemy1_controller.init = function(sprite) {
     sprite.animations.add('throb', null, 10, true);
     sprite.play('throb');
-    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,800), game.rnd.integerInRange(0,450), 100);
+    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,GAME_WIDTH), game.rnd.integerInRange(0,GAME_HEIGHT), 100);
 }
 
 var enemy2_controller = Object.create(enemy_controller);
 
 enemy2_controller.type = 'enemy2';
 enemy2_controller.point_value = 2;
+enemy2_controller.health = 2;
 enemy2_controller.spritelist = ['enemy5'];
 enemy2_controller.rotation = 15;
 enemy2_controller.init = function(sprite) {
     sprite.animations.add('throb', null, 10, true);
     sprite.play('throb');
-    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,800), game.rnd.integerInRange(0,450), 100);
+    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,GAME_WIDTH), game.rnd.integerInRange(0,GAME_HEIGHT), 100);
 };
 
 enemy2_controller.behavior = function(sprite) {
     if (game.rnd.integerInRange(0, 100) < 40) {
         if (game.rnd.integerInRange(0, 100) > 50) {
-            game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,800), game.rnd.integerInRange(0,450), 100);    
+            game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,GAME_WIDTH), game.rnd.integerInRange(0,GAME_HEIGHT), 100);    
         } else {
             game.physics.arcade.moveToObject(sprite, main.player, 200);
         }
     }
 }
 
+var bigenemy_controller = Object.create(enemy_controller);
+bigenemy_controller.type = 'bigenemy';
+bigenemy_controller.point_value = 6;
+bigenemy_controller.rotation = 15;
+bigenemy_controller.spritelist = ['bigenemy'];
+bigenemy_controller.health = 6;
+
+bigenemy_controller.init = function(sprite) {
+    sprite.animations.add('throb', null, 10, true);
+    sprite.play('throb');
+    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,GAME_WIDTH), game.rnd.integerInRange(0,GAME_HEIGHT), 100);
+};
+
+bigenemy_controller.onkill = function(sprite) {
+    var spawnx = sprite.x;
+    var spawny = sprite.y;
+    main.create_enemy('splitdude', spawnx, spawny);
+    main.create_enemy('splitdude', spawnx, spawny);
+    main.create_enemy('splitdude', spawnx, spawny);
+    main.create_enemy('splitdude', spawnx, spawny);
+};
+
+var splitdude_controller = Object.create(enemy_controller);
+splitdude_controller.type = 'splitdude';
+splitdude_controller.spritelist = ['splitdude'];
+splitdude_controller.control_interval = 500;
+splitdude_controller.target_player = true;
+
+splitdude_controller.init = function(sprite) {
+    game.physics.arcade.moveToXY(sprite, game.rnd.integerInRange(0,GAME_WIDTH), game.rnd.integerInRange(0,GAME_HEIGHT), 200);
+    this.next_control = game.time.now + 200;
+};
+
+splitdude_controller.behavior = function(sprite) {
+    if (this.target_player) {
+        game.physics.arcade.moveToObject(sprite, main.player, 300);
+        this.target_player = false;
+    } else {
+        if (game.rnd.integerInRange(0, 100) < 10) {
+            this.target_player = true;
+        }
+    }
+};
+
 /*** End Enemy Definitions! ***/
+
+/*** Powerups! ***/
+var powerup = {
+    property: null,
+    increment: 0,
+    player_effect: null,
+    sprite: null
+};
+
+var shot_damage_powerup = {
+    property: 'shot_damage',
+    increment: 1,
+    player_effect: 'player_effect_blue',
+    sprite: 'bluearrow'
+};
+
+var max_speed_powerup = {
+    property: 'max_player_speed',
+    increment: 5,
+    player_effect: 'player_effect_yellow',
+    sprite: 'yellowarrow'
+};
+
+var accel_powerup = {
+    property: 'player_acceleration',
+    increment: 1,
+    player_effect: 'player_effect_green',
+    sprite: 'greenarrow'
+};
+
+var powerup_types = [
+    shot_damage_powerup,
+    max_speed_powerup,
+    accel_powerup
+];
 
 var spawn_thresholds = {
     0: {
@@ -96,20 +176,23 @@ var spawn_thresholds = {
         enemies: ['enemy1', 'animated', 'curver']
     },
     50: {
-        enemies: ['animated', 'curver']
+        enemies: ['animated', 'curver', 'bigenemy']
     }
 };
 
 var enemy_types = {
     'enemy1': enemy1_controller,
     'animated': enemy2_controller,
-    'curver': curver_controller
+    'curver': curver_controller,
+    'bigenemy': bigenemy_controller,
+    'splitdude': splitdude_controller
 };
 
 var main = {
     spawn_chance: 5,
     kills: 0,
     score: 0,
+    shot_damage: 1,
     max_player_speed: 120,
     player_acceleration: 5,
     shot_cooldown: 165,
@@ -191,7 +274,55 @@ var main = {
             this.next_shot = game.time.now + this.shot_cooldown;
         }
     },
-    create_enemy: function() {
+    create_powerup: function(type, x ,y) {
+        var powerup = game.add.sprite(x, y, type.sprite);
+        game.physics.enable(powerup, Phaser.Physics.ARCADE);
+        powerup.controller = type;
+        this.powerups.add(powerup);
+    },
+    random_powerup: function() {
+        var type = powerup_types[game.rnd.integerInRange(0,powerup_types.length-1)];
+        this.create_powerup(type, game.rnd.integerInRange(0, GAME_WIDTH), game.rnd.integerInRange(0, GAME_HEIGHT));    
+    },
+    create_enemy: function(type, x, y) {
+        var _this = this;
+        // do check recycle in this.enemies_by_type
+        if (this.enemies_by_type[type] !== undefined && this.enemies_by_type[type].countDead() > 0) {
+            var enemy = this.enemies_by_type[type].getFirstDead();
+            enemy.reset(x, y, enemy.controller.health);
+        } else {   
+            if (this.enemies_by_type[type] === undefined) {
+                this.enemies_by_type[type] = game.add.group();
+            }
+            var controller = Object.create(enemy_types[type]);
+            var enemy = game.add.sprite(x, y, controller.spritelist[game.rnd.integerInRange(0, controller.spritelist.length-1)]);
+            game.physics.enable(enemy, Phaser.Physics.ARCADE);
+            enemy.anchor.setTo(controller.centerx, controller.centery);
+            enemy.health = controller.health;
+            enemy.body.collideWorldBounds = true;
+                
+            enemy.controller = controller;
+            
+            this.enemies_by_type[type].add(enemy);
+            
+            enemy.events.onKilled.add(function() {
+                controller.onkill(enemy);
+                _this.kills += 1;
+                _this.score += controller.point_value;
+            });
+        }
+        
+        if (game.physics.arcade.overlap(enemy, this.player)) {
+            enemy.x = game.rnd.integerInRange(0, GAME_WIDTH);
+            enemy.y = game.rnd.integerInRange(0, GAME_HEIGHT);
+        }
+        
+        enemy.controller.init(enemy);
+        enemy.controller.behavior(enemy);
+        enemy.controller.next_control = game.time.now + enemy.controller.control_interval;
+    },
+    random_spawn: function() {
+        var _this = this;
         var current_level = 0;
         var thresholds = Object.keys(spawn_thresholds);
         for (var i = 0; i < thresholds.length; i++) {
@@ -203,31 +334,9 @@ var main = {
         
         
         var new_enemy_type = current_level.enemies[game.rnd.integerInRange(0, current_level.enemies.length-1)];
-       
-        // do check recycle in this.enemies_by_type
-        if (this.enemies_by_type[new_enemy_type] !== undefined && this.enemies_by_type[new_enemy_type].countDead() > 0) {
-            var enemy = this.enemies_by_type[new_enemy_type].getFirstDead();
-            enemy.reset(game.rnd.integerInRange(0,800), game.rnd.integerInRange(0,450));
-        } else {   
-            if (this.enemies_by_type[new_enemy_type] === undefined) {
-                this.enemies_by_type[new_enemy_type] = game.add.group();
-            }
-            var controller = Object.create(enemy_types[new_enemy_type]);
-            var enemy = game.add.sprite(game.rnd.integerInRange(0,800), game.rnd.integerInRange(0,450), controller.spritelist[game.rnd.integerInRange(0, controller.spritelist.length-1)]);
-            game.physics.enable(enemy, Phaser.Physics.ARCADE);
-            enemy.anchor.setTo(0.5, 0.5);
-            enemy.body.collideWorldBounds = true;
-                
-            enemy.controller = controller;
-            
-            this.enemies_by_type[new_enemy_type].add(enemy);
-            
-        }
-        
-        
-        enemy.controller.init(enemy);
-        enemy.controller.behavior(enemy);
-        enemy.controller.next_control = game.time.now + enemy.controller.control_interval;
+        var x = game.rnd.integerInRange(0, GAME_WIDTH);
+        var y = game.rnd.integerInRange(0, GAME_HEIGHT);
+        this.create_enemy(new_enemy_type, x, y);
     },
     explode_at: function(x, y) {
         
@@ -269,14 +378,21 @@ var main = {
         game.load.image('yellowparticle', 'assets/yellowparticle.png');
         game.load.image('orangeparticle', 'assets/orangeparticle.png');
         
-        game.load.image('enemy3', 'assets/enemy3.png');
-        game.load.image('enemy4', 'assets/enemy4.png');
+        game.load.image('bluearrow', 'assets/bluearrow.png');
+        game.load.image('greenarrow', 'assets/greenarrow.png');
+        game.load.image('yellowarrow', 'assets/yellowarrow.png');
+        game.load.image('player_effect_blue', 'assets/bluespriteeffect.png');
+        game.load.image('player_effect_green', 'assets/greenspriteeffect.png');
+        game.load.image('player_effect_yellow', 'assets/yellowspriteeffect.png');
+        
+        game.load.image('splitdude','assets/split_dude.png');
+        
         
         game.load.spritesheet('enemy1', 'assets/enemy1_new.png', 16, 16);
         game.load.spritesheet('enemy2', 'assets/enemy2_new.png', 16, 16);
-        
         game.load.spritesheet('enemy5', 'assets/enemy_sheet.png', 20, 20);
         game.load.spritesheet('enemy6', 'assets/greensheet.png', 10, 10);
+        game.load.spritesheet('bigenemy', 'assets/bigenemy_sheet.png', 35, 35);
         
         game.load.image('touch_circle','assets/control_circle.png');
     },
@@ -292,6 +408,10 @@ var main = {
             this.joystick_right = Object.create(touch_joystick);
             this.joystick_right.init(game, {x: 650, y: 300}, 'touch_circle');
         }
+        
+        for (var i = 0; i < powerup_types.length; i++) {
+            this[powerup_types[i].property+'_default'] = this[powerup_types[i].property];
+        }
 
         // Entity groups
         this.projectiles = game.add.group();
@@ -299,6 +419,10 @@ var main = {
         
         // Emitter groups
         this.explosions = game.add.group();
+        
+        this.powerups = game.add.group();
+        this.player_effects = game.add.group();
+        this.current_effect_offset = 0;
         
         game.physics.startSystem(Phaser.Physics.ARCADE);
  
@@ -329,6 +453,8 @@ var main = {
         
         this.projectiles.destroy(true);
         this.explosions.destroy(true);
+        this.powerups.destroy(true);
+        this.player_effects.destroy(true);
         
         highscores.update(this.score);
         highscores.render_scores(game);
@@ -346,6 +472,10 @@ var main = {
         
     },
     reset: function() {
+        for (var i = 0; i < powerup_types.length; i++) {
+            this[powerup_types[i].property] = this[powerup_types[i].property+'_default'];
+        }
+        
         highscores.clear();
         this.is_game_over = false;
         
@@ -353,6 +483,8 @@ var main = {
         
         this.projectiles = game.add.group();
         this.explosions = game.add.group();
+        this.powerups = game.add.group();
+        this.player_effects = game.add.group();
         
         this.kills = 0;
         this.score = 0;
@@ -434,22 +566,30 @@ var main = {
         var frames = frame_delta / (1000 / 60); // Pretending 60fps
         
         this.player.angle += 15 * frames;
+        
+        this.player_effects.forEach(function(effect) {
+            effect.x = _this.player.x;
+            effect.y = _this.player.y;
+            effect.angle += 20 * frames;
+        });
+        
         Object.keys(this.enemies_by_type).forEach(function(type) {
             // Collide stuff
             game.physics.arcade.collide(_this.projectiles, _this.enemies_by_type[type], function(projectile, enemy) {
                 projectile.kill();
                 _this.explode_at(enemy.x, enemy.y);
-                _this.score += enemy.controller.point_value;
-                enemy.kill();
-                _this.kills += 1;
+                
+                enemy.damage(_this.shot_damage);
             });
             
             game.physics.arcade.collide(_this.player, _this.enemies_by_type[type], function(player, enemy) {
-                _this.player.kill();
-                _this.explode_at(_this.player.x, _this.player.y);
-                game.time.events.add(1000, function() {
-                    _this.game_over();
-                });
+                if (_this.player.alive) {
+                    _this.player.kill();
+                    _this.explode_at(_this.player.x, _this.player.y);
+                    game.time.events.add(1000, function() {
+                        _this.game_over();
+                    });
+                }
             });
             
             _this.enemies_by_type[type].forEachAlive(function(enemy) {
@@ -463,11 +603,27 @@ var main = {
             
            
         });
+        
+        game.physics.arcade.collide(this.player, this.powerups, function(player, powerup) {
+            powerup.kill();
+            var new_effect = game.add.sprite(player.x, player.y, powerup.controller.player_effect);
+            console.log(new_effect);
+            new_effect.angle = _this.current_effect_offset;
+            new_effect.anchor.setTo(0.5, 0.5);
+            //console.log(new_effect);
+            _this[powerup.controller.property] += powerup.controller.increment;
+            _this.player_effects.add(new_effect);
+            this.current_effect_offset += 21;
+        });
  
         // Generate Enemies
         
         if (game.rnd.integerInRange(0, 1000) < (this.spawn_chance * frames)) {
-            this.create_enemy();
+            this.random_spawn();
+        }
+        
+        if (game.rnd.integerInRange(0, 10000) < 5) {
+            this.random_powerup();
         }
         
         this.spawn_chance = 5 + (Math.floor(this.score / 10));
@@ -484,7 +640,7 @@ window.onload = function() {
         main.go_fullscreen();
     };
 
-    game = new Phaser.Game(800, 450, Phaser.AUTO, 'gameDiv');
+    game = new Phaser.Game(GAME_WIDTH, GAME_HEIGHT, Phaser.AUTO, 'gameDiv');
     game.state.add('main', main);
     game.state.start('main');
 }
